backend/requirements.txt
asgiref
Babel
certifi
charset-normalizer
Django
django-cors-headers
django-environ
django-extensions
django-filter
django-jet-reboot
django-money
djangorestframework
gunicorn
idna
Markdown
packaging
psycopg
psycopg-binary
psycopg-pool
#py-moneyed
python-slugify
pytz
requests
simplejson
sqlparse
text-unidecode
typing_extensions
urllib3
